+++ config/options.js
@@ -1,4 +1,0 @@
-global.boundaries = {
-    top: 900,
-    left: 1500
-}