+++ middleware/request_log.js
@@ -1,0 +1,12 @@
+module.exports = function(req, res, next){
+    let excluded = false;
+    logging.excluded.forEach((fileType) => {
+        if(req.url.endsWith(fileType)){
+            excluded = true;
+        }
+    });
+    if(!excluded){
+        Logger.logRequest(req);
+    }
+    next();
+}